refactor: Participant playrer/dealer 공통로직 분리
